perf: implement synchronous remote request execution
Introduce the ability to execute 'API Requests' 'Synchronously', 
bypassing the 'Job Queue' when immediate 'Responses' are required, 
including refactoring the 'Request Processing' logic to use a 
centralized 'Remote Execution' handler and improving 'Metadata 
Extraction'.

- Add synchronous execution mode for API requests.
- Bypass job queue for time-critical operations.
- Refactor request processing to use centralized remote handler.
- Improve metadata extraction from API responses.
